% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_flow.R
\name{calc_Rig_grad}
\alias{calc_Rig_grad}
\title{Calculate the richardson gradient from the model output}
\usage{
calc_Rig_grad(df_harm)
}
\arguments{
\item{df_harm}{model meteo parameteres}
}
\description{
calculates the Richardson gradient number around hub height (>50 and <145).
}
