% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_flow.R
\name{direction_cor}
\alias{direction_cor}
\title{Directions correlation at different height levels}
\usage{
direction_cor(df_harm, df_dir)
}
\description{
calculates the correlation coefficient between the modeled and observed directions based on the cor.circular
function for the circular package. The Hlev are interpolated to the observed height levels.
}
