% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_self_similarity.R
\name{get_complete_days}
\alias{get_complete_days}
\title{Get complete days}
\usage{
get_complete_days(df_h, n_obs = 144)
}
\arguments{
\item{df_h}{dataframe with columns time an u. time as POSIXct and u as numeric in m/s.}

\item{n_obs}{number of observations per day, set to 144 (complete day with 10 minute intervals
which is the standard interval of wind lidar data).}
}
\description{
returns a dataframe with only complete observation days.
}
\author{
Marieke Dirksen
}
