% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lidar_locations.R
\name{get_lidar_locations}
\alias{get_lidar_locations}
\title{Get spatial lidar information}
\usage{
get_lidar_locations()
}
\description{
Combines the coordinates of the lidar stations into an sf object.
}
\author{
Marieke Dirksen
}
