% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_self_similarity.R
\name{get_mean_S}
\alias{get_mean_S}
\title{Get the mean daily structures}
\usage{
get_mean_S(df_h)
}
\arguments{
\item{df_h}{dataframe with columns time an u. time as POSIXct and u as numeric in m/s.}
}
\description{
Applies the function \link{StructureFunction_NaN} to a dataframe with columns
time and u. The structures are calculated from normalized u for each complete day.
}
\author{
Marieke Dirksen
}
