% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vars.R
\name{kinematic_heat_flux}
\alias{kinematic_heat_flux}
\title{Kinematic potential temperature flux}
\usage{
kinematic_heat_flux(Ta, p, reh, H, LE)
}
\arguments{
\item{Ta}{actual temperature in K}

\item{p}{pressure in kPa}

\item{reh}{relative humidity in \\%}

\item{H}{is variable hfss is WINS50}

\item{LE}{is variable hfls_eva+hfls_sbl}
}
\description{
Calculates the kinematic potential temperature flux from the sensible and latent heat fluxes, using
equation 2.33, p.12, from the Thesis of Coen Hennipman (TUD).
}
