% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vars.R
\name{mixing_ratio}
\alias{mixing_ratio}
\title{Mixing ratio}
\usage{
mixing_ratio(Ta, p, reh)
}
\arguments{
\item{Ta}{actual temperature}

\item{p}{pressure}

\item{reh}{relative humidity}
}
\description{
Mixing ratio
}
