% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vars.R
\name{obukhov}
\alias{obukhov}
\title{Obukhov length}
\usage{
obukhov(z, Ta, p, reh, tauu, tauv, H, LE)
}
\arguments{
\item{z}{height}

\item{Ta}{actual temperature}

\item{p}{pressure}

\item{reh}{relative humidity}

\item{tauu}{stress in the u dir}

\item{tauv}{stress in the v dir}

\item{H}{sensible heat}

\item{LE}{latent heat}

\item{khf}{kinematic heat flux}
}
\description{
calculate the Obukhov length
}
