% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_rose_plot_routine.R
\name{plot.windrose}
\alias{plot.windrose}
\title{Wind rose plotting routine}
\usage{
\method{plot}{windrose}(
  data,
  spd,
  dir,
  spdres = 2,
  dirres = 30,
  spdmin = 2,
  spdmax = 20,
  spdseq = NULL,
  palette = "YlGnBu",
  countmax = NA,
  debug = 0
)
}
\arguments{
\item{data}{input data with column names spd and dir}

\item{spd}{numerical values of the speed}

\item{dir}{direction}

\item{spdres}{slipt number}

\item{dirres}{resolution of the number of directions}

\item{spdmin}{minimum speed}

\item{spdmax}{maximum speed}

\item{spdseq}{NULL}

\item{palette}{color palette to be used}

\item{countmax}{NA}

\item{debug}{to debug the code}
}
\description{
Based on ggplot2 plotting routine
}
