% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ESS.R
\name{plot_ESSh}
\alias{plot_ESSh}
\title{Plot similarity's}
\usage{
plot_ESSh(SR, df, h_levels = 11, main)
}
\arguments{
\item{SR}{similarities as output from \link{get_mean_S}.}

\item{df}{data input from \link{get_mean_S} to add statistics to the plot.}

\item{h_levels}{number of height levels}

\item{main}{header of the plot}
}
\description{
Plots S3 vs S2 and S2 vs S4.
}
\author{
Marieke Dirksen
}
