% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OSM.R
\name{plot_OSM_wind_meels}
\alias{plot_OSM_wind_meels}
\title{Plot the Wind Meels for the research domain}
\usage{
plot_OSM_wind_meels(update_farms = FALSE)
}
\arguments{
\item{update_farms}{TRUE or FALSE. If TRUE a new dataset from OSM is downloaded, if FALSE
the rds file in the data folder is loaded.}
}
\value{
Returns a ggplot2 object which can be modified and cropped for a smaller domain.
}
\description{
The boundaries of the domain are set to -5W, 50S, 10E, 57N.
Additional data from natural earth was added for the plotting routine.
}
\author{
Marieke Dirksen
}
