% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RS.R
\name{plot_RS}
\alias{plot_RS}
\title{Plot structure}
\usage{
plot_RS(SR, df, main)
}
\arguments{
\item{SR}{similarities as output from \link{get_mean_S}.}

\item{df}{data input from \link{get_mean_S} to add statistics to the plot.}

\item{main}{header of the plot}
}
\description{
Plots R versus S.
}
\author{
Marieke Dirksen
}
