% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RS.R
\name{plot_RSh}
\alias{plot_RSh}
\title{Plot structure for different heights}
\usage{
plot_RSh(SR, df, h_levels = 11, main)
}
\arguments{
\item{SR}{similarities as output from \link{get_mean_S}.}

\item{df}{data input from \link{get_mean_S} to add statistics to the plot.}

\item{h_levels}{number of individual height levels, expects SR to have a seperate column with h}

\item{main}{header of the plot}
}
\description{
Plots R versus S for different heights.
}
\author{
Marieke Dirksen
}
