% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_curve.R
\name{power_curve}
\alias{power_curve}
\title{Power curve}
\usage{
power_curve(Ui)
}
\arguments{
\item{Ui}{velocities for which the power is estimated}
}
\value{
relative power production at each Ui
Last updated: October 10, 2018
modified by Marieke on November 19, 2020 to R-code
Following J. King (2014)
}
\description{
simplified version of the power curve
}
