% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ANH.R
\name{read_ANH}
\alias{read_ANH}
\title{Read the wind data from ANH}
\usage{
read_ANH(dir = "D:/data/Lidar/ANH/", h = 60, what = "Speed")
}
\arguments{
\item{dir}{directory of the files (allowed to be stored in subfolders).}

\item{h}{height of the measurement (40m,60m,76m,80m,100m,116m,160m,200m,250m,290m).}

\item{what}{choose between Speed and Direction. Wind speed as sqrt(u^2+v^2).}
}
\description{
Function to read the wind speed and direction from the ANH lidar. Data can be
requested from \url{https://offshorewind.rvo.nl/windwaternh}.
}
\author{
Marieke Dirksen
}
