% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Daarle.R
\name{read_BSA}
\alias{read_BSA}
\title{Read the wind data from Borselle Alpha}
\usage{
read_BSA(
  dir = "D:/data/Lidar/Steven/",
  h = 50,
  what = "Speed",
  stn = "Borssele_Alpha",
  period = "JA2019"
)
}
\arguments{
\item{dir}{directory of the files (allowed to be stored in subfolders).}

\item{h}{height of the measurement, depends on period:
For July and August 2019: 59m,84m,112m,121m,143m,168m,194m,205m,225m,245m or 295m
For September 2019: 60m,84m,109m,118m,140m,165m,191m,202m,225m,245m or 295m.}

\item{what}{choose between Speed and Direction. Wind speed as sqrt(u^2+v^2).}

\item{stn}{The stations of Steven (function not included in online version), stn name is Borssele_Alpha.}

\item{period}{JA2019 for July and August 2019 or S for September 2019.}
}
\description{
Function to read the wind speed and direction from the Borselle Alpha lidar.
Dataset from KNMI, contact Steven for details.
}
\author{
Marieke Dirksen
}
