% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_BWFZ.R
\name{read_BWFZ}
\alias{read_BWFZ}
\title{Read the wind speed from BWFZ}
\usage{
read_BWFZ(dir, h = 40, what = "Speed")
}
\arguments{
\item{dir}{Directory where the wind files are stored.}

\item{h}{measurement height including 4m,30m, 40m, 60m, 80m, 100m, 120m, 140m, 160m, 180m and 200m.}

\item{what}{choose between Speed and Direction to get the wind speed (horizontal sqrt(u^2+v^2)) or wind direction.}
}
\description{
Reads the wind speed and direction from BWFZ at an existing measurement height. The data can be
downloaded from \url{https://offshorewind.rvo.nl/studiesborssele}. Note that the non-configurable reference
height of 40m is ignored during the height selection.
}
\author{
Marieke Dirksen
}
