% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Daarle.R
\name{read_Daarle}
\alias{read_Daarle}
\title{Read the wind data from Daarle}
\usage{
read_Daarle(
  dir = "D:/data/Lidar/Steven/",
  h = 50,
  what = "Speed",
  stn = "Daarle"
)
}
\arguments{
\item{dir}{directory of the files (allowed to be stored in subfolders).}

\item{h}{height of the measurement (50m,65m,80m,100m,126m,135m,145m,155m,180m,200m,220m,235m).}

\item{what}{choose between Speed and Direction. Wind speed as sqrt(u^2+v^2).}

\item{stn}{The stations of Steven (function not included in online version), stn name is Daarle.}
}
\description{
Function to read the wind speed and direction from the Daarle lidar.
Data only available at KNMI, contact Steven for details.
}
\author{
Marieke Dirksen
}
