% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Daarle.R
\name{read_Eemshaven}
\alias{read_Eemshaven}
\title{Read the wind data from Eemshaven}
\usage{
read_Eemshaven(
  dir = "D:/data/Lidar/Steven/",
  h = 58,
  what = "Speed",
  stn = "Eemshaven"
)
}
\arguments{
\item{dir}{directory of the files (allowed to be stored in subfolders).}

\item{h}{height of the measurement (42m,58m,71m,84m,96m,109m,121m,133m,146m,158m,183m,200m).}

\item{what}{choose between Speed and Direction. Wind speed as sqrt(u^2+v^2).}

\item{stn}{The stations of Steven (function not included in online version), stn is Eemshaven.}
}
\description{
Function to read the wind speed and direction from the Eemshaven lidar.
Data only available at KNMI, contact Steven for details.
}
\author{
Marieke Dirksen
}
