% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EPL.R
\name{read_EPL}
\alias{read_EPL}
\title{Read the wind data from EPL}
\usage{
read_EPL(dir = "D:/data/Lidar/EPL/", h = 63, what = "Speed")
}
\arguments{
\item{dir}{directory of the files (allowed to be stored in subfolders).}

\item{h}{height of the measurement (63m,91m,116m,141m,166m,191m,216m,241m,266m,291m).}

\item{what}{choose between Speed and Direction. Wind speed as sqrt(u^2+v^2).}
}
\description{
Function to read the wind speed and direction from the EPL lidar. Data can be
requested from \url{https://www.windopzee.net/meet-locaties/europlatform/}.
}
\author{
Marieke Dirksen
}
