% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_FINO2.R
\name{read_FINO2}
\alias{read_FINO2}
\title{Read the wind data from FINO2}
\usage{
read_FINO2(dir = "D:/data/Lidar/FINO2/", h = 69, what = "Speed")
}
\arguments{
\item{dir}{directory of the files (allowed to be stored in subfolders).}

\item{h}{height of the measurement (63m,91m,116m,141m,166m,191m,216m,241m,266m,291m).}

\item{what}{choose between Speed and Direction. Wind speed as sqrt(u^2+v^2).}
}
\description{
Function to read the wind speed and direction from the FINO2 lidar. Data can be
requested from \url{https://orsted.com/en/our-business/offshore-wind/wind-data}.
}
\author{
Marieke Dirksen
}
