% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nc.R
\name{read_flux_nc}
\alias{read_flux_nc}
\title{Load the fluxes of the nc file}
\usage{
read_flux_nc(nc_file)
}
\arguments{
\item{nc_file}{name and full path to the file}
}
\value{
a data frame with obs
}
\description{
Load the fluxes of the nc file
}
