% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HKZ.R
\name{read_HKZ}
\alias{read_HKZ}
\title{Read the wind data from HKZ}
\usage{
read_HKZ(dir = "D:/data/Lidar/HKZ", h = 60, what = "Speed", stn = "HKZA")
}
\arguments{
\item{dir}{directory of the files (allowed to be stored in subfolders).}

\item{h}{height of the measurement (4m,30m,40m,60m,80m,100m,120m,140m,160m,180m,200m).}

\item{what}{choose between Speed and Direction. Wind speed as sqrt(u^2+v^2).}

\item{stn}{choose between HKZA and HKZB lidar locations.}
}
\description{
Function to read the wind speed and direction from the HKZ lidar. Data can be
downloaded from \url{https://www.windopzee.net/meet-locaties/hollandse-kust-zuid-hkz/data/}.
}
\author{
Marieke Dirksen
}
