% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_K13a.R
\name{read_K13A}
\alias{read_K13A}
\title{Read the wind data from HKW}
\usage{
read_K13A(dir = "D:/data/Lidar/K13/", h = 60, what = "Speed")
}
\arguments{
\item{dir}{directory of the files (allowed to be stored in subfolders).}

\item{h}{height of the measurement (63m,91m,116m,141m,166m,191m,216m,241m,266m,291m).}

\item{what}{choose between Speed and Direction. Wind speed as sqrt(u^2+v^2).}
}
\description{
Function to read the wind speed and direction from the HKZ lidar. Data can be
downloaded from \url{https://www.windopzee.net/meet-locaties/hollandse-kust-zuid-hkz/data/}.
}
\author{
Marieke Dirksen
}
