% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_TNW.R
\name{read_TNW}
\alias{read_TNW}
\title{Read the wind data from TNW}
\usage{
read_TNW(dir = "D:/data/Lidar/TNW", h = 60, what = "Speed", stn = "TNWA")
}
\arguments{
\item{dir}{directory of the files (allowed to be stored in subfolders).}

\item{h}{height of the measurement (4m, 30m ,40m ,60m ,80m ,100m ,120m ,140m ,160m ,180m ,200m ,250m).}

\item{what}{choose between Speed and Direction. Wind speed as sqrt(u^2+v^2).}

\item{stn}{choose between TNWA and TNWB lidar locations.}
}
\description{
Function to read the wind speed and direction from the TNW lidar. Data can be
downloaded from \url{https://offshorewind.rvo.nl/TNW_WindAndWater}.
}
\author{
Marieke Dirksen
}
