% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vars.R
\name{Ri_grad}
\alias{Ri_grad}
\title{Richardson Gradient number}
\usage{
Ri_grad(thetav, u, v, z, g = 9.81)
}
\arguments{
\item{thetav}{virtual potential temperature}

\item{u}{wind speed eastward}

\item{v}{wind speed southward}

\item{z}{height levels corresponding to the model/measurement levels of thetav,u,v.}

\item{g}{gravitational constant}
}
\description{
Linear approximation of the richardson gradient number for a height difference z
}
