% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAR_raster.R
\name{SAR_raster}
\alias{SAR_raster}
\title{Raster from SAR}
\usage{
SAR_raster(nc_file, land_mask = europe_sf, r_res = 500)
}
\arguments{
\item{nc_file}{nc file from SAR data}

\item{land_mask}{sf polygon to mask the land areas (for example the naturalearth dataset)}

\item{r_res}{resolution of the raster grid in meters}
}
\description{
Create a regular lat lon grid from the SAR points data
}
\author{
Marieke Dirksen
}
