% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_self_similarity.R
\name{StructureFunction_NaN}
\alias{StructureFunction_NaN}
\title{Structures for extended self similarity analysis}
\usage{
StructureFunction_NaN(x, DT = 6)
}
\arguments{
\item{x}{data from a timeseries, for example wind measurements}

\item{DT}{Number of hours over which the similarity structures are caculated (only holds if the timeinterval equals 10 minutes).}
}
\description{
This function calculates the structures for extended self similarity of timeseries.
}
\author{
Marieke Dirksen, written for R from matlab code from Sukanta Basu
}
