% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_flow.R
\name{subset_flow}
\alias{subset_flow}
\title{Subset turbulent and laminar flow}
\usage{
subset_flow(I, df_level, df_harm)
}
\arguments{
\item{I}{times with tubulent/laminar flow in POSIXct}

\item{df_level}{observations from the lidar}

\item{df_harm}{model input}
}
\description{
Subset turbulent and laminar flow based on the time indices I. The model input will be interpolated to
the model levels and for each timestamp the errors will be calculated.
}
