% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vars.R
\name{theta}
\alias{theta}
\title{Potential temperature}
\usage{
theta(Ta, p)
}
\arguments{
\item{Ta}{actual temperature in K}

\item{p}{pressure in kPa}
}
\value{
theta in K
}
\description{
calculates the potential temperature with standard values for cp, cv and p0.
}
