% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vars.R
\name{thetav}
\alias{thetav}
\title{Virtual Potential temperature}
\usage{
thetav(Ta, p, reh)
}
\arguments{
\item{Ta}{actual temperature in K}

\item{p}{pressure in kPa}

\item{reh}{relative humidity in \\%}
}
\value{
theta in K
}
\description{
calculates the potential temperature with standard values for cp, cv and p0.
}
