% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vars.R
\name{u_friction}
\alias{u_friction}
\title{Friction velocity}
\usage{
u_friction(tauu, tauv)
}
\arguments{
\item{tauu}{eastward stress component}

\item{tauv}{southern stress component}
}
\value{
uf friction velocity m/s
}
\description{
calculates the friction velocity assuming a constant air density at the surface
}
