% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vars.R
\name{virtual_temperature}
\alias{virtual_temperature}
\title{Virtual temperature}
\usage{
virtual_temperature(Ta, p, reh)
}
\arguments{
\item{Ta}{actual temperature}

\item{p}{pressure}

\item{reh}{relative humidity}
}
\description{
Virtual temperature
}
