% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_meels}
\alias{wind_meels}
\title{Wind Meels data from OSM}
\format{
An object of class \code{list} (inherits from \code{osmdata}, \code{osmdata_sf}) of length 8.
}
\usage{
wind_meels
}
\description{
Dataset with wind meels from OSM. The wind meel data is stored in
the osm features "generator:source" and "wind".
}
\author{
Marieke Dirksen
}
\keyword{datasets}
